requirements.txt
pandas
scikit-learn
fastapi
uvicorn
joblib
sqlalchemy
psycopg2-binary
python-dotenv
